+++ command.h
@@ -4,8 +4,8 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-void command_receive(uint8_t *rx_data, uint8_t rx_size, void (*handler)(uint8_t));
+void command_receive(uint8_t *rx_data, uint16_t rx_size, void (*handler)(uint8_t));
 
-void command_send(uint8_t *tx_data, uint8_t tx_size, void (*handler)(uint8_t));
+void command_send(uint8_t *tx_data, uint16_t tx_size, void (*handler)(uint8_t));
 
 #endif /* COMMAND_H */
